make vehicleInssurance api

+++ cancleRide.php
@@ -15,7 +15,7 @@ function sendPushNotification($driverId)
     {
         $deviceToken = $token['deviceToken'];
         $serverKey = 'AAAAzpUqMlE:APA91bEXySQ-4aw7rQB6Sloy2WLgyAr4XIEToPk5xo98u-wDOICMTC1ExzysY0SYBBio24gHaFgQlPh0BV3RIL-Ls34Y-d-_v205s79Bxj6MZ-tH2WI7_mlp6jGXtsxB5gNmloxmIIgQ'; // Replace with your Firebase Server Key
-        $data = [
+        $row = [
             'to' => $deviceToken, // The recipient's FCM token
             'notification' => [
                 'title' => 'Safe Drive',
@@ -31,7 +31,7 @@ function sendPushNotification($driverId)
         curl_setopt($ch, CURLOPT_HTTPHEADER, $headers);
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false); // Not recommended for production
-        curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($data));
+        curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($row));
         $response = curl_exec($ch);
         if ($response === false) {
             die('Error: ' . curl_error($ch));
@@ -44,6 +44,7 @@ function sendPushNotification($driverId)
 if(isset($_POST['userId']))
 {
     $userId = $_POST['userId'];
+    $cancleReason = $_POST['cancelReason'];
 
     $checkRideQuery = "SELECT * FROM book_ride WHERE userId = '$userId'";
     $checkRide = mysqli_query($con,$checkRideQuery);
@@ -51,19 +52,34 @@ if(isset($_POST['userId']))
     {
         while($row = mysqli_fetch_assoc($checkRide))
         {
-            $driverId = $row['driverId'];
-            // $response['driverId'] = $driverId;/
-            $deleteUserRideQuery = "DELETE FROM book_ride WHERE userId = '$userId'";
-            $deleteUserRide = mysqli_query($con,$deleteUserRideQuery);
+            // $driverId = $row['driverId'];
+            // // $response['driverId'] = $driverId;/
+            // $deleteUserRideQuery = "DELETE FROM book_ride WHERE userId = '$userId'";
+            // $deleteUserRide = mysqli_query($con,$deleteUserRideQuery);
 
-            if($deleteUserRide)
+            // if($deleteUserRide)
+            // {
+            //     $deleteRequestQuery = "DELETE from request where user_id = '$userId'";
+            //     $deleteRequest = mysqli_query($con,$deleteRequestQuery);
+            //     sendPushNotification($driverId);
+
+            //     $response['status'] = "200";
+            //     $response['message'] = "cancle your ride";
+            // }
+
+            $insertCompleteQuery = "INSERT INTO completerides(userId,driverId,pessangerName,pickup_letitude,pickup_longitude,drop_letitude,drop_longitude,vehicle_type,amount,payment_mode,rideStatus,cancleReason,fromAddress,toAddress,booking_date)VALUES('{$row['userId']}','{$row['driverId']}','{$row['pessangerName']}','{$row['pickup_letitude']}','{$row['pickup_longitude']}','{$row['drop_letitude']}','{$row['drop_longitude']}','{$row['vehicle_type']}','{$row['amount']}','{$row['payment_mode']}','cancel','$cancleReason','{$row['fromAddress']}','{$row['toAddress']}','{$row['booking_date']}')";
+            $insertComplete = mysqli_query($con,$insertCompleteQuery);  
+
+            if($insertComplete)
             {
-                $deleteRequestQuery = "DELETE from request where user_id = '$userId'";
+                $deleteUserRideQuery = "DELETE FROM book_ride WHERE userId = '$userId'";
+                $deleteUserRide = mysqli_query($con,$deleteUserRideQuery);
+
+                $deleteRequestQuery = "DELETE from request where user_id = '$userId' and driverId = '$driverId'";
                 $deleteRequest = mysqli_query($con,$deleteRequestQuery);
-                sendPushNotification($driverId);
 
                 $response['status'] = "200";
-                $response['message'] = "cancle your ride";
+                $response['message'] = "your Ride cancel";
             }
         }
     }
@@ -73,18 +89,41 @@ elseif(isset($_POST['userId']) && isset($_POST['driverId']))
 {
     $userId = $_POST['userId'];
     $driverId = $_POST['driverId'];
- 
+    $cancleReason = $_POST['cancelReason'];
     
-    $deleteRideQuery = "DELETE from book_ride where userId = '$userId' and driverId = '$driverId'";
-    $deleteRide = mysqli_query($con,$deleteRideQuery);
+    // $deleteRideQuery = "DELETE from book_ride where userId = '$userId' and driverId = '$driverId'";
+    // $deleteRide = mysqli_query($con,$deleteRideQuery);
 
-    if ($deleteRide) {
-        $deleteRequestQuery = "DELETE from request where user_id = '$userId' and driverId = '$driverId'";
-        $deleteRequest = mysqli_query($con,$deleteRequestQuery);
+    // if ($deleteRide) {
+    //     $deleteRequestQuery = "DELETE from request where user_id = '$userId' and driverId = '$driverId'";
+    //     $deleteRequest = mysqli_query($con,$deleteRequestQuery);
 
-        sendPushNotification($userId);
-        $response['status'] = "200";
-        $response['message'] = "cancle the ride";
+    //     sendPushNotification($userId);
+    //     $response['status'] = "200";
+    //     $response['message'] = "cancle the ride";
+    // }
+
+    $checkRideQuery = "SELECT * FROM book_ride WHERE userId = '$userId' AND driverId = '$driverId'";
+    $checkRide = mysqli_query($con,$checkRideQuery);
+    if(mysqli_num_rows($checkRide) > 0)
+    {
+        while($row = mysqli_fetch_assoc($checkRide))
+        {
+            $insertCompleteQuery = "INSERT INTO completerides(userId,driverId,pessangerName,pickup_letitude,pickup_longitude,drop_letitude,drop_longitude,vehicle_type,amount,payment_mode,rideStatus,cancleReason,fromAddress,toAddress,booking_date)VALUES('{$row['userId']}','{$row['driverId']}','{$row['pessangerName']}','{$row['pickup_letitude']}','{$row['pickup_longitude']}','{$row['drop_letitude']}','{$row['drop_longitude']}','{$row['vehicle_type']}','{$row['amount']}','{$row['payment_mode']}','cancel','$cancleReason','{$row['fromAddress']}','{$row['toAddress']}','{$row['booking_date']}')";
+            $insertComplete = mysqli_query($con,$insertCompleteQuery);  
+
+            if($insertComplete)
+            {
+                $deleteUserRideQuery = "DELETE FROM book_ride WHERE userId = '$userId' AND driverId = '$driverId'";
+                $deleteUserRide = mysqli_query($con,$deleteUserRideQuery);
+
+                $deleteRequestQuery = "DELETE from request where user_id = '$userId' and driverId = '$driverId'";
+                $deleteRequest = mysqli_query($con,$deleteRequestQuery);
+
+                $response['status'] = "200";
+                $response['message'] = "your Ride cancel";
+            }
+        }
     }
 }
 else
